requirements.txt
ultralytics
gdown
supervision
numpy
opencv-python
transformers
umap-learn
scikit-learn
tqdm
sentencepiece
protobuf
